+++ bindings/c/examples/basic.c
@@ -25,10 +25,10 @@ int main()
 {
     /* Initialize a operator for "memory" backend, with no options */
     opendal_result_operator_new result = opendal_operator_new("memory", 0);
-    assert(result.operator_ptr != NULL);
+    assert(result.op != NULL);
     assert(result.error == NULL);
 
-    opendal_operator* op = result.operator_ptr;
+    opendal_operator* op = result.op;
 
     /* Prepare some data to be written */
     opendal_bytes data = {
