+++ bindings/c/examples/error_handle.c
@@ -26,10 +26,10 @@ int main()
 {
     /* Initialize a operator for "memory" backend, with no options */
     opendal_result_operator_new result = opendal_operator_new("memory", 0);
-    assert(result.operator_ptr != NULL);
+    assert(result.op != NULL);
     assert(result.error == NULL);
 
-    opendal_operator* op = result.operator_ptr;
+    opendal_operator* op = result.op;
 
     /* The read is supposed to fail */
     opendal_result_read r = opendal_operator_read(op, "/testpath");
