+++ bindings/c/tests/bdd.cpp
@@ -44,7 +44,7 @@ protected:
         opendal_result_operator_new result = opendal_operator_new(scheme.c_str(), options);
         EXPECT_TRUE(result.error == nullptr);
 
-        this->p = result.operator_ptr;
+        this->p = result.op;
         EXPECT_TRUE(this->p);
 
         opendal_operator_options_free(options);
